commit - parsing data from websocket added

+++ src/main/java/sep4/iot/gateway/model/SensorEntry.java
@@ -7,7 +7,7 @@ public class SensorEntry implements Serializable {
     private int entry_key;
     private int user_key;
     private String hweui;
-    private int entry_time;
+    private long entry_time;
 
     private float air_temperature;
     private float air_humidity;
@@ -22,7 +22,7 @@ public class SensorEntry implements Serializable {
     public SensorEntry() {
     }
 
-    public SensorEntry(int entry_key, String hweui, int user_key, int entry_time, float air_temperature, float air_humidity, int air_co2, float light_level, float desired_air_temperature, float desired_air_humidity, int desired_air_co2, float desired_light_level) {
+    public SensorEntry(int entry_key, String hweui, int user_key, long entry_time, float air_temperature, float air_humidity, int air_co2, float light_level, float desired_air_temperature, float desired_air_humidity, int desired_air_co2, float desired_light_level) {
         this.entry_key = entry_key;
         this.user_key = user_key;
         this.hweui = hweui;
@@ -37,7 +37,7 @@ public class SensorEntry implements Serializable {
         this.desired_light_level = desired_light_level;
     }
 
-    public SensorEntry(int entry_key, int user_key, String hweui, int entry_time, float air_temperature, float air_humidity, int air_co2, float light_level) {
+    public SensorEntry(int entry_key, int user_key, String hweui, long entry_time, float air_temperature, float air_humidity, int air_co2, float light_level) {
         this.entry_key = entry_key;
         this.user_key = user_key;
         this.hweui = hweui;
@@ -110,11 +110,11 @@ public class SensorEntry implements Serializable {
         this.user_key = user_key;
     }
 
-    public int getEntry_time() {
+    public long getEntry_time() {
         return entry_time;
     }
 
-    public void setEntry_time(int entry_time) {
+    public void setEntry_time(long entry_time) {
         this.entry_time = entry_time;
     }
